Share function to get save folder and prepare forced self updates
This allows the launcher to get the Isaac save folder without requiring
a full UI initialization, as it may be necessary for the launcher to
check the state of the lock file before anything has been setup.

This also prepares the launcher to initiate a self update from any
context, in case users need to forcibly update to fixed a botched
installation.