Self-updater management
This adds base logic for checking the availability of updates, and
adds stronger error handling for everything self-updater related.

WaitForInputIdle() on the self-updater now uses a timeout to prevent
the launcher from freezing in case there is an issue with the
self-updater. The view part of the launcher now uses a retry mechanism
in case the self-updater times out.