I want to finish this thing pleeeeeeeeeeeeeeease
* Split the install / update logic from MainFrame into the new
InstallationManager class. This makes it easier to work on either
the GUI or the logic.
* Removed the option to select where to install Repentogon, this
makes things easier for everyone. Repentogon's DLLs will always
be installed next to isaac-ng.exe.
* Removed the option to let the user select where to install the
configuration file of the launcher, it is now installed in the
Repentance save folder.
* Removed the distinction between general launcher config and
configuration of the launcher's options when it comes to INI file.
All parameters are now unified under a single ini file.
* Added the ability to select unstable updates of Repentogon.
* Improved feedback during the download of Repentogon / launcher
update by only logging a transfer once per 100ms instead of each
time cURL receives something.
* Fixed an issue with the Sha256 function where it could fail due
to the comparison between hashes being case sensitive.
* Other fixes and improvements (FromSoftware level of changelog)