Better updates support + fix static boxes
Summary: This improves support for unstable updates of the launcher,
improves the GUI when it comes to forced updates and fixes the
layout of static boxes.

* Unstable updates of the launcher are now managed through a
preprocessor define. This requires a bit more work depending on what
we settle on regarding how upgrades should be managed.
* The GUI has been reworked so that the Repentogon forced updates
buttons are now part of the Advanced Options box. This box has been
moved to a separate modal to reduce the cluttering of the main GUI.
* Static boxes have been reworked to use wxStaticBoxSizer instead
of traditional sizers. This resolves many issues with the general
placement and size of static boxes.