Finalize the injection process
Summary: This completes the injection process in the following
ways:
* The call to CreateProcess no longer expects isaac-ng.exe to be
present in the current folder. Instead it takes the full path as
indicated by the user in the window.
* The call to CreateProcess properly defines the working environment
of isaac-ng.exe instead of using the same environment as the launcher.
* Remove the code of the injector part of the launcher from the
trampoline section and keep only the actual injected function in
that section.
* The console boolean is now injected alongside the pointers towards
the kernel32.dll functions, removing the need to use --console in
the CLI.
* Fixed the call to VirtualProtectEx that sets the injected page to
RX (from RWX). The issue came from an improper call to
VirtualProtectEx, where the address was NULL instead of the base of
the page (now properly acquired through VirtualQueryEx).
* Updated the "zhlLauncher.dll" string to "zhlLoader.dll" to be in
line with the changes in Repentogon.