Extracting the archive and starting the launcher
Summary: This adds the final steps, i.e. extracting the content of the
archive and then starting the launcher, assuming extraction went well.
There is still a bug due to the fact that the updater loads the DLLs
that are to be replaced by the updater, so this still needs to be fixed.