Merge launcher and updater. Add unpacker.
Summary: This changes how the launcher manages its self update, and
solves the issue where the updater was unable to replace some files
because it was using them.

The updater is now part of the launcher and downloads the release. It
decompresses the archive into a single binary file. If then starts the
embedded unpacker executable, that synchronizes with the launcher, kills
it, then unpacks the content of the binary file.

The unpacker has no external dependencies and as such is not affected by
changes in any of the DLLs or other files.