Fix forgotten corner case to install Repentogon
If the user provided a valid Isaac installation and configuration file,
but the configuration file did not specify that Repentogon should be
updated automatically, then the launcher would not install Repentogon.

This fixes the issue by checking for the validity of the Repentogon
installation if the wizard is skipped and a valid Isaac installation is
found regardless.